fix(fe2): fsevents error resolved + storybook updated + .env.example updated to be more reasonable (#1715)
* fix: resolved fsevents issue + ensure-tailwind-deps building even when not needed

* fixed storybook

* upgrading storybook

* .env.example update
