+++ packages/frontend-2/.env.example
@@ -4,7 +4,7 @@ PORT=8081
 NUXT_PUBLIC_LOG_LEVEL=debug
 NUXT_PUBLIC_LOG_PRETTY=true
 
-NUXT_PUBLIC_API_ORIGIN=http://localhost:3000
+NUXT_PUBLIC_API_ORIGIN=http://127.0.0.1:3000
 
 NUXT_PUBLIC_MIXPANEL_TOKEN_ID=acd87c5a50b56df91a795e999812a3a4
 NUXT_PUBLIC_MIXPANEL_API_HOST=https://analytics.speckle.systems
