+++ packages/frontend-2/nuxt.config.ts
@@ -49,12 +49,7 @@ export default defineNuxtConfig({
   alias: {
     // Rewriting all lodash calls to lodash-es for proper tree-shaking & chunk splitting
     lodash: 'lodash-es',
-    '@vue/apollo-composable': '@speckle/vue-apollo-composable',
-    // We need browser polyfills for crypto & zlib cause they seem to be bundled for the web
-    // for some reason when running the dev server or storybook. Doesn't appear that these
-    // actually appear in any client-side bundles tho!
-    crypto: require.resolve('rollup-plugin-node-builtins'),
-    zlib: require.resolve('browserify-zlib')
+    '@vue/apollo-composable': '@speckle/vue-apollo-composable'
   },
 
   vite: {
