+++ packages/frontend-2/package.json
@@ -82,16 +82,16 @@
     "@nuxt/devtools": "^0.2.5",
     "@nuxtjs/tailwindcss": "^6.3.0",
     "@speckle/tailwind-theme": "workspace:^",
-    "@storybook/addon-a11y": "^7.0.0-beta.54",
-    "@storybook/addon-docs": "^7.0.0-beta.54",
-    "@storybook/addon-essentials": "^7.0.0-beta.54",
-    "@storybook/addon-interactions": "^7.0.0-beta.54",
-    "@storybook/addon-links": "^7.0.0-beta.54",
-    "@storybook/jest": "^0.0.11-next.0",
-    "@storybook/test-runner": "0.10.0-next.8",
-    "@storybook/testing-library": "^0.0.14-next.1",
-    "@storybook/vue3": "^7.0.0-beta.54",
-    "@storybook/vue3-vite": "^7.0.0-beta.54",
+    "@storybook/addon-a11y": "^7.1.0",
+    "@storybook/addon-docs": "^7.1.0",
+    "@storybook/addon-essentials": "^7.1.0",
+    "@storybook/addon-interactions": "^7.1.0",
+    "@storybook/addon-links": "^7.1.0",
+    "@storybook/jest": "^0.1.0",
+    "@storybook/test-runner": "0.11.0",
+    "@storybook/testing-library": "^0.2.0",
+    "@storybook/vue3": "^7.1.0",
+    "@storybook/vue3-vite": "^7.1.0",
     "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/line-clamp": "^0.4.2",
     "@testing-library/vue": "^6.6.1",
@@ -111,7 +111,7 @@
     "eslint": "^8.24.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-nuxt": "^4.0.0",
-    "eslint-plugin-storybook": "^0.6.6",
+    "eslint-plugin-storybook": "^0.6.13",
     "eslint-plugin-vue": "^9.5.1",
     "eslint-plugin-vuejs-accessibility": "^1.2.0",
     "jest": "27",
@@ -125,7 +125,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "rollup-plugin-node-builtins": "^2.1.2",
-    "storybook": "^7.0.0-beta.54",
+    "storybook": "^7.1.0",
     "stylelint": "^15.10.1",
     "stylelint-config-prettier": "^9.0.3",
     "stylelint-config-recommended-vue": "^1.4.0",
